added func to separate users - finished testing.

+++ src/main/java/com/diary/db/EntryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EntryRepository extends MongoRepository<Entry, Long> {
+public interface EntryRepository extends MongoRepository<Entry, Integer> {
 }
